images/requirements.txt
flask==0.12
python-magic==0.4.13
